crypto_api/requirements.txt
fastapi
uvicorn
pycryptodome
cryptography
python-multipart
